+++ backend/src/test/java/com/hiddenartist/backend/global/DataInitializerTest.java
@@ -7,7 +7,7 @@ import com.hiddenartist.backend.domain.artist.persistence.Artist;
 import com.hiddenartist.backend.domain.artist.persistence.repository.ArtistRepository;
 import com.hiddenartist.backend.domain.artwork.persistence.Artwork;
 import com.hiddenartist.backend.domain.artwork.persistence.repository.ArtworkRepository;
-import com.hiddenartist.backend.global.config.CustomDataJpaTest;
+import com.hiddenartist.backend.global.config.AbstractMySQLRepositoryTest;
 import com.hiddenartist.backend.global.config.TestDataInitializer;
 import com.hiddenartist.backend.global.type.EntityToken;
 import java.util.List;
@@ -15,8 +15,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
-@CustomDataJpaTest
-public class DataInitializerTest {
+public class DataInitializerTest extends AbstractMySQLRepositoryTest {
 
   @Autowired
   TestDataInitializer initializer;
