--- backend/src/test/java/com/hiddenartist/backend/global/config/CustomDataJpaTest.java
@@ -1,18 +1,0 @@
-package com.hiddenartist.backend.global.config;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import org.springframework.context.annotation.Import;
-import org.springframework.transaction.annotation.Transactional;
-
-@Target({ElementType.TYPE})
-@Retention(RetentionPolicy.RUNTIME)
-@DataJpaTest
-@Transactional
-@Import({TestConfig.class, TestContainerSetup.class})
-public @interface CustomDataJpaTest {
-
-}
